Extract auth funcionality to auth-hook
