+++ src/shared/hooks/auth-hook.js
@@ -1,0 +1,59 @@
+import { useState, useEffect, useCallback } from "react";
+
+let logoutTimer;
+
+export const useAuth = () => {
+    const [userId, setUserId] = useState(false);
+  const [tokenExpirationDate, setTokenExpirationDate] = useState();
+  const [token, setToken] = useState(null);
+
+  const login = useCallback((uid, token, expirationDate) => {
+    setToken(token);
+    const currentExpirationDate =
+      expirationDate || new Date(new Date().getTime() + 1000 * 60 * 60);
+    setTokenExpirationDate(currentExpirationDate);
+    localStorage.setItem(
+      "userData",
+      JSON.stringify({
+        userId: uid,
+        token: token,
+        expiration: currentExpirationDate.toISOString(),
+      })
+    );
+    setUserId(uid);
+  }, []);
+
+  const logout = useCallback(() => {
+    setToken(null);
+    setTokenExpirationDate(null);
+    setUserId(null);
+    localStorage.removeItem("userData");
+  }, []);
+
+  useEffect(() => {
+    if (token && tokenExpirationDate) {
+      const remainingTimeInMiliseconds = tokenExpirationDate.getTime() - new Date().getTime();
+      logoutTimer = setTimeout(logout, remainingTimeInMiliseconds);
+    } else {
+      clearTimeout(logoutTimer);
+    }
+  }, [token, logout, tokenExpirationDate]);
+
+  useEffect(() => {
+    const storedData = JSON.parse(localStorage.getItem("userData"));
+    if (
+      storedData &&
+      storedData.token &&
+      new Date(storedData.expiration) > new Date()
+    ) {
+      login(
+        storedData.userId,
+        storedData.token,
+        new Date(storedData.expiration)
+      );
+    }
+  }, [login]);
+
+
+  return { token, login, logout, userId };
+}
